InitSQL separated
