+++ app/src/main/java/com/shubhamgupta16/simplewallpaper/activities/SplashActivity.java
@@ -40,8 +40,6 @@ public class SplashActivity extends AppCompatActivity {
         initInterstitial();
 
         sqlHelper = new SQLHelper(this);
-        setupWallpapers();
-        setupCategories();
 
 
     }
@@ -82,60 +80,5 @@ public class SplashActivity extends AppCompatActivity {
                 showInterstitial();
             }
         });
-    }
-
-    private void setupCategories() {
-        try {
-            JSONArray array = new JSONArray(readJSONFromAsset("categories"));
-            for (int i = 0; i < array.length(); i++) {
-                JSONObject object = array.getJSONObject(i);
-                CategoryPOJO pojo = new CategoryPOJO(
-                        object.getString("name"),
-                        object.getString("preview1"),
-                        object.getString("preview2"),
-                        object.getString("preview3"),
-                        0);
-                sqlHelper.insertCategory(pojo);
-            }
-        } catch (JSONException e) {
-            e.printStackTrace();
-        }
-    }
-
-    private void setupWallpapers() {
-        try {
-            JSONArray array = new JSONArray(readJSONFromAsset("wallpapers"));
-            for (int i = 0; i < array.length(); i++) {
-                JSONObject object = array.getJSONObject(i);
-                WallsPOJO pojo = new WallsPOJO(
-                        0,
-                        object.getString("name"),
-                        object.getString("previewUrl"),
-                        object.getString("url"),
-                        object.getString("categories"),
-                        object.optBoolean("premium", false)
-                );
-                sqlHelper.insertWallpaper(pojo);
-            }
-        } catch (JSONException e) {
-            e.printStackTrace();
-        }
-    }
-
-    @SuppressWarnings("ResultOfMethodCallIgnored")
-    private String readJSONFromAsset(String fileName) {
-        String json;
-        try {
-            InputStream is = getAssets().open(fileName + ".json");
-            int size = is.available();
-            byte[] buffer = new byte[size];
-            is.read(buffer);
-            is.close();
-            json = new String(buffer, StandardCharsets.UTF_8);
-        } catch (IOException ex) {
-            ex.printStackTrace();
-            return null;
-        }
-        return json;
     }
 }
