+++ app/src/main/res/layout/activity_more_walls.xml
@@ -29,6 +29,6 @@
         android:id="@+id/adView"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        ads:adSize="SMART_BANNER"
+        ads:adSize="BANNER"
         ads:adUnitId="@string/main_banner_id"/>
 </RelativeLayout>
